Split apart CLI files
